pyproject.toml
[tool.pytest.ini_options]
addopts = "--html=.local/test_report/report.html --cov --cov-config pyproject.toml --cov-report term --cov-report html --cov-report xml --junitxml .local/test_report/test_report.xml --typeguard-packages=btctoy"
junit_family = "xunit2"
cache_dir = ".local/pytest_cache"
asyncio_mode = "strict"

[tool.coverage.run]
branch = true
source = ["btctoy"]
data_file = ".local/.coverage"

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "if 0:",
  "if False:",
]

[tool.coverage.html]
directory = ".local/test_report/coverage_html"

[tool.coverage.xml]
output = ".local/test_report/coverage.xml"

[tool.isort]
# This configuration is compatible with black
# https://black.readthedocs.io/en/stable/compatible_configs.html#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 1                   # Force multi-line imports, to reduce merge conflict probability
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.poetry]
name = "tp-programming-bitcoin-transactions"
version = "0.1.0"
description = ""
authors = ["Celeborn2BeAlive <laurent.noel.c2ba@gmail.com>"]
readme = "README.md"
packages = [{ include = "btctoy", from = "src" }]

[tool.poetry.scripts]
btctoy = "btctoy.__main__:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
poetry-version = "^0.2.0"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
debugpy = "^1.7.0"
flake8 = "^6.1.0"
flake8-absolute-import = "^1.0.0.1"
flake8-annotations = "^3.0.1"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.7.10"
flake8-builtins = "^2.1.0"
flake8-comprehensions = "^3.14.0"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.0.0"
flake8-pytest-style = "^1.7.2"
pep8-naming = "^0.13.3"
mypy = "^1.5.1"
pytest = "^7.4.2"
typeguard = "^4.1.4"
pytest-html = "^4.0.1"
pytest-cov = "^4.1.0"
nox = "^2023.4.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
